frontend/requirements.txt
streamlit==1.41.0
httpx==0.27.0
python-dotenv==1.0.1
streamlit-autorefresh==1.0.1
